remove: ➖ GeoLocation Provider
BREAKING CHANGE: 🧨 BREAKING CHANGE!
