+++ src/context/index.tsx
@@ -1,6 +1,5 @@
 import Modals from './Modals'
 import Sidebar from './Sidebar'
 import Toast from './Toast'
-import GeoLocation from './GeoLocation'
 
-export { Modals, Sidebar, Toast, GeoLocation }
+export { Modals, Sidebar, Toast }
