+++ src/pages/_app.tsx
@@ -4,7 +4,7 @@ import type { AppProps } from 'next/app'
 import { SessionProvider } from 'next-auth/react'
 
 // Context
-import { Modals, Sidebar, Toast, GeoLocation } from '@/context'
+import { Modals, Sidebar, Toast } from '@/context'
 import { Session } from 'next-auth'
 
 if (process.env.NODE_ENV === 'development') {
@@ -17,13 +17,11 @@ function App({ Component, pageProps }: AppProps<{ session: Session }>) {
   return (
     <SessionProvider session={pageProps.session}>
       <Toast.ToastProvider>
-        <GeoLocation.GeoLocationProvider>
-          <Modals.ModalsProvider>
-            <Sidebar.SidebarProvider>
-              <Component {...pageProps} />
-            </Sidebar.SidebarProvider>
-          </Modals.ModalsProvider>
-        </GeoLocation.GeoLocationProvider>
+        <Modals.ModalsProvider>
+          <Sidebar.SidebarProvider>
+            <Component {...pageProps} />
+          </Sidebar.SidebarProvider>
+        </Modals.ModalsProvider>
       </Toast.ToastProvider>
     </SessionProvider>
   )
